# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for IndiLingo - learn Indian regional languages with alphabet/script lessons
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new learner profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a learner profile with progress stats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /exercises/{exerciseId}/mistake:
    post:
      operationId: recordMistake
      tags: [review]
      summary: Record a missed exercise for spaced repetition
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAction"
      responses:
        "200":
          description: OK
  /exercises/{exerciseId}/master:
    post:
      operationId: masterExercise
      tags: [review]
      summary: Mark an exercise as mastered (remove from review queue)
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAction"
      responses:
        "200":
          description: OK
  /users/{userId}/review:
    get:
      operationId: getReviewExercises
      tags: [review]
      summary: Get exercises queued for review (spaced repetition)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /users/{userId}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get detailed progress stats for a user (lessons completed per language, streaks, XP)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
components:
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: string
        lessonId:
          type: string
        type:
          type: string
          description: "Exercise type: multiple_choice, translate, match_pairs, fill_blank, script_practice"
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        romanization:
          type: ["string", "null"]
          description: "Romanized pronunciation guide (used in script_practice and translate exercises)"
        nativeScript:
          type: ["string", "null"]
          description: "The native script character or word being taught"
        order:
          type: integer
      required:
        - id
        - lessonId
        - type
        - prompt
        - options
        - correctAnswer
        - order
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        xp:
          type: integer
        streak:
          type: integer
        hearts:
          type: integer
        lastActiveDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - xp
        - streak
        - hearts
        - lastActiveDate
        - createdAt
    UserInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UserStats:
      type: object
      properties:
        userId:
          type: string
        totalLessonsCompleted:
          type: integer
        totalXp:
          type: integer
        streak:
          type: integer
        languageProgress:
          type: array
          items:
            $ref: "#/components/schemas/LanguageProgress"
        badges:
          type: array
          items:
            type: string
      required:
        - userId
        - totalLessonsCompleted
        - totalXp
        - streak
        - languageProgress
        - badges
    LanguageProgress:
      type: object
      properties:
        languageId:
          type: string
        languageName:
          type: string
        lessonsCompleted:
          type: integer
        totalLessons:
          type: integer
      required:
        - languageId
        - languageName
        - lessonsCompleted
        - totalLessons
    ReviewAction:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
...